+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adding `SliderWidget`, added `SliderOrientation` for horizontal or vertical orientation.
 - Added `slider` example, added `SliderHorizontal` orientation for testing.
+- Added mouse scrolling functionality to slider.
 
 ## 0.4.17